moved test_agentset_* to its backend folder
