+++ tests/pandas/test_agentset_pandas.py
